car-deprecated {car}	R Documentation

Deprecated Functions in the car Package

Description

These functions are provided for compatibility with older versions of the car package only, and may be removed eventually. Commands that worked in versions of the car package prior to version 3.0-0 will not necessarily work in version 3.0-0 and beyond, or may not work in the same manner.

Usage

bootCase(...)

nextBoot(...)

Arguments

...
arguments to pass to methods.

Details

These functions are replaced by Boot.

See Also

See Also Boot

[Package car version 3.1-1 Index]