car-package {car}	R Documentation

Companion to Applied Regression

Description

Functions and Datasets to Accompany J. Fox and S. Weisberg, An R Companion to Applied Regression, Second Edition, Sage, 2011.

Details

Package: 	 car

Version: 	 2.1-3

Date: 	 2016-07-23

Depends: 	 R (>= 3.2.0)

Imports: 	 MASS, mgcv, nnet, pbkrtest (>= 0.3-2), quantreg, grDevices, utils, stats, graphics

Suggests: 	 alr4, boot, leaps, lme4, lmtest, Matrix, MatrixModels, mgcv, nlme, rgl (>= 0.93.960), sandwich, SparseM, survival, survey

License: 	 GPL (>= 2)

URL: 	 http://CRAN.R-project.org/package=car, http://socserv.socsci.mcmaster.ca/jfox/Books/Companion, https://r-forge.r-project.org/projects/car/

Author(s)

John Fox <jfox@mcmaster.ca> and Sanford Weisberg. We are grateful to Douglas Bates, Gabriel Baud-Bovy, David Firth, Michael Friendly, Gregor Gorjanc, Spencer Graves, Richard Heiberger, Rafael Laboissiere, Georges Monette, Henric Nilsson, Derek Ogle, Brian Ripley, Achim Zeleis, and R Core for various suggestions and contributions.

Maintainer: John Fox <jfox@mcmaster.ca>

[Package car version 2.1-3 Index]