carWeb {car}	R Documentation

Access to the R Companion to Applied Regression Website

Description

This function will access the website for An R Companion to Applied Regression.

Usage

carWeb(...)

Arguments

...
optional arguments if present are passed to utils::browseURL

Value

Displays the web page for An R Companion to Applied Regression, Third Edition, Sage

[Package car version 3.1-5 Index]