logit {car}	R Documentation

Logit Transformation

Description

Compute the logit transformation of proportions or percentages.

Usage

logit(p, percents=range.p[2] > 1, adjust)

Arguments

p
numeric vector or array of proportions or percentages.

percents
TRUE for percentages.

adjust
adjustment factor to avoid proportions of 0 or 1; defaults to 0 if there are no such proportions in the data, and to .025 if there are.

Details

Computes the logit transformation logit = log[p/(1 - p)] for the proportion p.

If p = 0 or 1, then the logit is undefined. logit can remap the proportions to the interval (adjust, 1 - adjust) prior to the transformation. If it adjusts the data automatically, logit will print a warning message.

Value

a numeric vector or array of the same shape and size as p.

Author(s)

John Fox jfox@mcmaster.ca

References

Fox, J. and Weisberg, S. (2019) An R Companion to Applied Regression, Third Edition, Sage.

See Also

probabilityAxis

Examples

options(digits=4)
logit(.1*0:10)
##  [1] -3.6636 -1.9924 -1.2950 -0.8001 -0.3847  0.0000  0.3847
##  [8]  0.8001  1.2950  1.9924  3.6636
## Warning message: 
## proportions remapped to (0.025, 0.975) in: logit(0.1 * 0:10) 

logit(.1*0:10, adjust=0)
##  [1]    -Inf -2.1972 -1.3863 -0.8473 -0.4055  0.0000  0.4055
##  [8]  0.8473  1.3863  2.1972     Inf

[Package car version 3.0-0 Index]