plot.powerTransform {car}	R Documentation

plot Method for powerTransform Objects

Description

This function provides a simple function for plotting data using power transformations.

Usage

## S3 method for class 'powerTransform'
plot(x, z = NULL, round = TRUE, plot = pairs, ...)

Arguments

x
name of the power transformation object

z
Additional variables of the same length as those used to get the transformation to be plotted, default is NULL.

round
If TRUE, the default, use rounded transforms, if FALSE use the MLEs.

plot
Plotting method. Default is pairs. Another possible choice is scatterplot.matrix from the car package.

...
Optional arguments passed to the plotting method

Details

The data used to estimate transformations using powerTransform are plotted in the transformed scale.

Value

None. Produces a graph as a side-effect.

Author(s)

Sanford Weisberg, <sandy@umn.edu>

References

Weisberg, S. (2014) Applied Linear Regression, Fourth Edition, Wiley.

Fox, J. and Weisberg, S. (2011) An R Companion to Applied Linear Regression, Second Edition, Sage.

See Also

powerTransform

Examples

summary(a3 <- powerTransform(cbind(len, ADT, trks, sigs1) ~ hwy, Highway1))
with(Highway1, plot(a3, z=rate, col=as.numeric(hwy)))

[Package car version 2.1-3 Index]