car-deprecated {car}	R Documentation

Deprecated Functions in car Package

Description

These functions are provided for compatibility with older versions of the car package only, and may be removed eventually. Commands that worked in versions of the car package prior to version 2.0-0 will not necessarily work in version 2.0-0 and beyond, or may not work in the same manner.

Usage

av.plot(...)
av.plots(...)
box.cox(...)
bc(...)
box.cox.powers(...)
box.cox.var(...)
box.tidwell(...)
cookd(...)
confidence.ellipse(...)
ceres.plot(...)
ceres.plots(...)
cr.plot(...)
cr.plots(...)
data.ellipse(...)
durbin.watson(...)
levene.test(...)
leverage.plot(...)
leverage.plots(...)
linear.hypothesis(...)
ncv.test(...)
outlier.test(...)
qq.plot(...)
scatterplot.matrix(...)
spread.level.plot(...)

Arguments

...
pass arguments down.

Details

av.plot and av.plots are now synonyms for the avPlot and avPlots functions.

box.cox and bc are now synonyms for bcPower.

box.cox.powers is now a synonym for powerTransform.

box.cox.var is now a synonym for boxCoxVariable.

box.tidwell is now a synonym for boxTidwell.

cookd is now a synonym for cooks.distance in the stats package.

confidence.ellipse is now a synonym for confidenceEllipse.

ceres.plot and ceres.plots are now synonyms for the ceresPlot and ceresPlots functions.

cr.plot and cr.plots are now synonyms for the crPlot and crPlots functions.

data.ellipse is now a synonym for dataEllipse.

durbin.watson is now a synonym for durbinWatsonTest.

levene.test is now a synonym for leveneTest function.

leverage.plot and leverage.plots are now synonyms for the leveragePlot and leveragePlots functions.

linear.hypothesis is now a synonym for the linearHypothesis function.

ncv.test is now a synonym for ncvTest.

outlier.test is now a synonym for outlierTest.

qq.plot is now a synonym for qqPlot.

scatterplot.matrix is now a synonym for scatterplotMatrix.

spread.level.plot is now a synonym for spreadLevelPlot.

[Package car version 2.1-3 Index]