carHexsticker {car}	R Documentation

View the Official Hex Sticker for the car Package

Description

Open the official hex sticker for the car package in your browser

Usage

carHexsticker()

Value

Used for its side effect of openning the hex sticker for the car package in your browser.

Author(s)

John Fox jfox@mcmaster.ca

Examples

## Not run:  # intended for interactive use
carHexsticker()

## End(Not run)

[Package car version 3.1-2 Index]