carWeb {car}	R Documentation

Access to the R Companion to Applied Regression website

Description

This function will access the website for An R Companion to Applied Regression.

Usage

carWeb(page = c("webpage", "errata", "taskviews"), script, data)

Arguments

page
A character string indicating what page to open. The default "webpage" will open the main web page, "errata" displays the errata sheet for the book, and "taskviews" fetches and displays a list of available task views from CRAN.

script
The quoted name of a chapter in An R Companion to Applied Regression, like "chap-1", "chap-2", up to "chap-8". All the R commands used in that chapter will be displayed in your browser, where you can save them as a text file.

data
The quoted name of a data file in An R Companion to Applied Regression, like "Duncan.txt" or "Prestige.txt". The file will be opened in your web browser. You do not need to specify the extension .txt

Value

Either a web page or a PDF document is displayed. Only one of the three arguments page, rfile, or data, should be used.

Author(s)

Sanford Weisberg, based on the function UsingR in the UsingR package by John Verzani

References

Fox, J. and Weisberg, S. (2011) An R Companion to Applied Regression, Second Edition, Sage.

Examples

## Not run: carWeb()

[Package car version 2.1-3 Index]