crPlots {car}	R Documentation

Component+Residual (Partial Residual) Plots

Description

These functions construct component+residual plots (also called partial-residual plots) for linear and generalized linear models.

Usage

crPlots(model, terms = ~., layout = NULL, ask, main, 
    ...)

crp(...)

crPlot(model, ...)

## S3 method for class 'lm'
crPlot(model, variable, 
  id.method = list(abs(residuals(model, type="pearson")), "x"),
  labels, 
  id.n = if(id.method[1]=="identify") Inf else 0,
  id.cex=1, id.col=palette()[1], id.location="lr",
  order=1, line=TRUE, smoother=loessLine,
	smoother.args=list(), smooth, span,
  col=palette()[1], col.lines=palette()[-1],
  xlab, ylab, pch=1, lwd=2, grid=TRUE, ...)

Arguments

model
model object produced by lm or glm.

terms
A one-sided formula that specifies a subset of the predictors. One component-plus-residual plot is drawn for each term. The default ~. is to plot against all numeric predictors. For example, the specification terms = ~ . - X3 would plot against all predictors except for X3. If this argument is a quoted name of one of the predictors, the component-plus-residual plot is drawn for that predictor only.

layout
If set to a value like c(1, 1) or c(4, 3), the layout of the graph will have this many rows and columns. If not set, the program will select an appropriate layout. If the number of graphs exceed nine, you must select the layout yourself, or you will get a maximum of nine per page. If layout=NA, the function does not set the layout and the user can use the par function to control the layout, for example to have plots from two models in the same graphics window.

ask
If TRUE, ask the user before drawing the next plot; if FALSE, the default, don't ask. This is relevant only if not all the graphs can be drawn in one window.

main
The title of the plot; if missing, one will be supplied.

...
crPlots passes these arguments to crPlot. crPlot passes them to plot.

variable
A quoted string giving the name of a variable for the horizontal axis

id.method,labels,id.n,id.cex,id.col,id.location
Arguments for the labelling of points. The default is id.n=0 for labeling no points. See showLabels for details of these arguments.

order
order of polynomial regression performed for predictor to be plotted; default 1.

line
TRUE to plot least-squares line.

smoother
Function to add a nonparametric smooth.

smoother.args
see ScatterplotSmoothers for available smooethers and arguments.

smooth, span
these arguments are included for backwards compatility: if smooth=TRUE then smoother is set to loessLine, and if span is specified, it is added to smoother.args.

col
color for points; the default is the first entry in the current color palette (see palette and par).

col.lines
a list of at least two colors. The first color is used for the ls line and the second color is used for the fitted lowess line. To use the same color for both, use, for example, col.lines=c("red", "red")

xlab,ylab
labels for the x and y axes, respectively. If not set appropriate labels are created by the function.

pch
plotting character for points; default is 1 (a circle, see par).

lwd
line width; default is 2 (see par).

grid
If TRUE, the default, a light-gray background grid is put on the graph

Details

The function intended for direct use is crPlots, for which crp is an abbreviation.

The model cannot contain interactions, but can contain factors. Parallel boxplots of the partial residuals are drawn for the levels of a factor.

Value

NULL. These functions are used for their side effect of producing plots.

Author(s)

John Fox jfox@mcmaster.ca

References

Cook, R. D. and Weisberg, S. (1999) Applied Regression, Including Computing and Graphics. Wiley.

Fox, J. (2008) Applied Regression Analysis and Generalized Linear Models, Second Edition. Sage.

Fox, J. and Weisberg, S. (2011) An R Companion to Applied Regression, Second Edition, Sage.

See Also

ceresPlots, avPlots

Examples

crPlots(m<-lm(prestige~income+education, data=Prestige)) 
# get only one plot
crPlots(m, terms=~ . - education)

crPlots(lm(prestige ~ log2(income) + education + poly(women,2), data=Prestige))

crPlots(glm(partic != "not.work" ~ hincome + children, 
  data=Womenlf, family=binomial))

[Package car version 2.1-3 Index]