infIndexPlot {car}	R Documentation

Influence Index Plot

Description

Provides index plots of Cook's distances, leverages, Studentized residuals, and outlier significance levels for a regression object.

Usage

infIndexPlot(model, ...)

influenceIndexPlot(model, ...)

## S3 method for class 'lm'
infIndexPlot(model,
     vars=c("Cook", "Studentized", "Bonf", "hat"), 
     main="Diagnostic Plots",
     labels, id.method = "y",
     id.n = if(id.method[1]=="identify") Inf else 0,
     id.cex=1, id.col=palette()[1], id.location="lr", grid=TRUE, ...)

Arguments

model
A regression object of class lm or glm.

vars
All the quantities listed in this argument are plotted. Use "Cook" for Cook's distances, "Studentized" for Studentized residuals, "Bonf" for Bonferroni p-values for an outlier test, and and "hat" for hat-values (or leverages). Capitalization is optional. All may be abbreviated by the first one or more letters.

main
main title for graph

id.method,labels,id.n,id.cex,id.col,id.location
Arguments for the labelling of points. The default is id.n=0 for labeling no points. See showLabels for details of these arguments.

grid
If TRUE, the default, a light-gray background grid is put on the graph

...
Arguments passed to plot

Value

Used for its side effect of producing a graph. Produces four index plots of Cook's distance, Studentized Residuals, the corresponding Bonferroni p-values for outlier tests, and leverages.

Author(s)

Sanford Weisberg, sandy@umn.edu

References

Cook, R. D. and Weisberg, S. (1999) Applied Regression, Including Computing and Graphics. Wiley.

Fox, J. (2008) Applied Regression Analysis and Generalized Linear Models, Second Edition. Sage.

Fox, J. and Weisberg, S. (2011) An R Companion to Applied Regression, Second Edition, Sage.

Weisberg, S. (2014) Applied Linear Regression, Fourth Edition, Wiley.

See Also

cooks.distance, rstudent, outlierTest, hatvalues

Examples

m1 <- lm(prestige ~ income + education + type, Duncan)
influenceIndexPlot(m1)

[Package car version 2.1-3 Index]