invResPlot {car}	R Documentation

Inverse Response Plots to Transform the Response

Description

For a lm model, draws an inverse.response plot with the response Y on the vertical axis and the fitted values Yhat on the horizontal axis. Uses nls to estimate lambda in the function Yhat = b0 + b1(Y)^(lambda). Adds the fitted curve to the plot. invResPlot is an alias for inverseResponsePlot.

Usage

inverseResponsePlot(model, lambda=c(-1,0,1), robust=FALSE, xlab=NULL, ...)

## S3 method for class 'lm'
inverseResponsePlot(model, lambda=c(-1,0,1), robust=FALSE, 
   xlab=NULL, labels=names(residuals(model)), ...)

invResPlot(model, ...)

Arguments

model
A lm regression object

lambda
A vector of values for lambda. A plot will be produced with curves corresponding to these lambdas and to the least squares estimate of lambda

xlab
The horizontal axis label. If NULL, it is constructed by the function.

labels
Case labels if labeling is turned on; see invTranPlot and showLabels for arguments.

robust
If TRUE, then estimation uses Huber M-estimates with the median absolute deviation to estimate scale and k= 1.345. The default is FALSE.

...
Other arguments passed to invTranPlot and then to plot.

Value

As a side effect, a plot is produced with the response on the horizontal axis and fitted values on the vertical axis. Several lines are added to be plot as the ols estimates of the regression of Yhat on Y^(lambda), interpreting lambda = 0 to be natural logarithms.

Numeric output is a list with elements

lambda
Estimate of transformation parameter for the response

RSS
The residual sum of squares at the minimum if robust=FALSE. If robust = TRUE, the value of Huber objective function is returned.

Author(s)

Sanford Weisberg, sandy@umn.edu

References

Fox, J. and Weisberg, S. (2011) An R Companion to Applied Regression, Second Edition, Sage.

Pendergast, L, and Sheather, S. (in press). On sensitivity of response plot estimation of a robust estimation approach. Scandinavian Journal of Statistics.

Weisberg, S. (2014) Applied Linear Regression, Fourth Edition, Wiley, Chapter 7.

See Also

invTranPlot, powerTransform, showLabels

Examples

m2 <- lm(rate ~ log(len) + log(ADT) + slim + shld + log(sigs1), Highway1)
invResPlot(m2)

[Package car version 2.1-3 Index]