panel.car {car}	R Documentation

Panel Function for Coplots

Description

a panel function for use with coplot that plots points, a lowess line, and a regression line.

Usage

panel.car(x, y, col, pch, cex=1, span=0.5, lwd=2, 
  reg.line=lm, lowess.line=TRUE, ...)

Arguments

x
vector giving horizontal coordinates.

y
vector giving vertical coordinates.

col
point color.

pch
plotting character for points.

cex
character expansion factor for points.

span
span for lowess smoother.

lwd
line width, default is 2.

reg.line
function to compute coefficients of regression line, or FALSE for no line.

lowess.line
if TRUE plot lowess smooth.

...
other arguments to pass to functions lines and regLine.

Value

NULL. This function is used for its side effect: producing a panel in a coplot.

Author(s)

John Fox jfox@mcmaster.ca

See Also

coplot, regLine

Examples

coplot(prestige ~ income|education, panel=panel.car, 
  col="red", data=Prestige)

[Package car version 3.1-2 Index]