some {car}	R Documentation

Sample a Few Elements of an Object

Description

Randomly select a few elements of an object, typically a data frame, matrix, vector, or list. If the object is a data frame or a matrix, then rows are sampled.

Usage

some(x, ...)

## S3 method for class 'data.frame'
some(x, n=10, cols=NULL, ...)

## S3 method for class 'matrix'
some(x, n=10, cols=NULL, ...)

## Default S3 method:
some(x, n=10, ...)

Arguments

x
the object to be sampled.

n
number of elements to sample.

cols
if NULL, use all columns, if a vector of column names or numbers, use only the columns indicated

...
arguments passed down.

Value

Sampled elements or rows.

Note

These functions are adapted from head and tail in the utils package.

Author(s)

John Fox jfox@mcmaster.ca

References

Fox, J. and Weisberg, S. (2019) An R Companion to Applied Regression, Third Edition, Sage.

See Also

head, tail, brief.

Examples

some(Duncan)
some(Duncan, cols=names(Duncan)[1:3])

[Package car version 3.1-2 Index]