whichNames {car}	R Documentation

Position of Row Names

Description

These functions return the indices of the supplied row names of a data frame or names of another object, such as a vector or list. whichNames is just an alias for which.names.

Usage

whichNames(names, object, ...)

which.names(names, object, ...)

## S3 method for class 'data.frame'
whichNames(names, object, ...)

## Default S3 method:
whichNames(names, object, ...)

Arguments

names
a name or character vector of names.

object
a data frame or an object with a names attribute.

...
not used.

Value

Returns the index or indices of names in row names of the data frame or names of an object of another class.

Author(s)

John Fox jfox@mcmaster.ca

References

Fox, J. and Weisberg, S. (2019) An R Companion to Applied Regression, Third Edition, Sage.

Examples

whichNames(c('minister', 'conductor'), Duncan)

[Package car version 3.1-2 Index]