which.names {car}	R Documentation

Position of Row Names

Description

These functions return the indices of row names in a data frame or a vector of names. whichNames is just an alias for which.names.

Usage

which.names(names, object)
whichNames(...)

Arguments

names
a name or character vector of names.

object
a data frame or character vector of (row) names.

...
arguments to be passed to which.names.

Value

Returns the index or indices of names within object.

Author(s)

John Fox jfox@mcmaster.ca

References

Fox, J. and Weisberg, S. (2011) An R Companion to Applied Regression, Second Edition, Sage.

Examples

which.names(c('minister', 'conductor'), Duncan)
## [1]  6 16

[Package car version 2.1-3 Index]